Add default restart setting for smart-restart task
If you wish to use `capistrano3-puma` for it's handy generators
but in combination with `foreman` for its export feature, it
clashes with the default hook that restarts puma for you.

A setting has been introduced that emulates the old behavior
by default, but you can now opt out of default restart.